requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
python-multipart
